Move public functions from cli.py to api.py
