+++ pyscaffold/cli.py
@@ -7,24 +7,11 @@ from __future__ import absolute_import, print_function
 import argparse
 import os.path
 import sys
-from datetime import date
 
 import pyscaffold
 
-from . import info, repo, shell, templates, utils
-from .api import Scaffold
-from .exceptions import (
-    DirectoryAlreadyExists,
-    DirectoryDoesNotExist,
-    GitNotConfigured,
-    GitNotInstalled,
-    InvalidIdentifier)
+from . import api, shell, templates, utils
 from .extensions import cookiecutter, django, pre_commit, tox, travis
-from .structure import (
-    add_namespace,
-    apply_update_rules,
-    create_structure,
-    make_structure)
 
 __author__ = "Florian Wilhelm"
 __copyright__ = "Blue Yonder"
@@ -103,7 +90,7 @@ def add_default_args(parser):
                         version='PyScaffold {ver}'.format(ver=version))
 
 
-def _add_external_generators(parser):
+def add_external_generators(parser):
     """Add Django and Cookiecutter in a way they cannot be called together."""
     group = parser.add_mutually_exclusive_group()
     cookiecutter.augment_cli(group)
@@ -128,7 +115,7 @@ def parse_args(args):
     cli_creators = [
         add_default_args,
         # Built-in extensions:
-        _add_external_generators,
+        add_external_generators,
         travis.augment_cli,
         pre_commit.augment_cli,
         tox.augment_cli]
@@ -157,141 +144,12 @@ def parse_args(args):
     return {k: v for k, v in opts.items() if v is not None}
 
 
-def _verify_git():
-    """Check if git is installed and able to provide the required information.
-    """
-    if not info.is_git_installed():
-        raise GitNotInstalled
-    if not info.is_git_configured():
-        raise GitNotConfigured
-
-
-def get_default_opts(project_name, **aux_opts):
-    """Creates default options using auxiliary options as keyword argument
-
-    Use this function if you want to use PyScaffold from another application
-    in order to generate an option dictionary that can than be passed to
-    :obj:`create_project`.
-
-    Args:
-        project_name (str): name of the project
-        **aux_opts: auxiliary options as keyword parameters
-
-    Returns:
-        dict: options with default values set
-
-    Raises:
-        :obj:`DirectoryDoesNot` : raised if PyScaffold is told to
-            update an inexistent directory
-    """
-
-    # This function uses information from git, so make sure it is available
-    _verify_git()
-
-    # Merge the default options generated by argparse
-    opts = parse_args([project_name])
-    # Remove inadvertent double definition of project_name
-    aux_opts.pop('project', None)
-    opts.update(aux_opts)
-    opts.setdefault('package', utils.make_valid_identifier(opts['project']))
-    opts.setdefault('author', info.username())
-    opts.setdefault('email', info.email())
-    opts.setdefault('release_date', date.today().strftime('%Y-%m-%d'))
-    opts.setdefault('year', date.today().year)
-    opts.setdefault('license', 'none')
-    opts.setdefault('description', 'Add a short description here!')
-    opts.setdefault('url', 'http://...')
-    opts.setdefault('version', pyscaffold.__version__)
-    opts.setdefault('title',
-                    '='*len(opts['project']) + '\n' + opts['project'] + '\n' +
-                    '='*len(opts['project']))
-    classifiers = ['Development Status :: 4 - Beta',
-                   'Programming Language :: Python']
-    opts.setdefault('classifiers', utils.list2str(
-        classifiers, indent=4, brackets=False, quotes=False, sep=''))
-    opts.setdefault('url', 'http://...')
-    # Initialize empty list of all requirements
-    opts.setdefault('requirements', list())
-    opts['namespace'] = utils.prepare_namespace(opts['namespace'])
-    if opts['namespace']:
-        opts['root_pkg'] = opts['namespace'][0]
-        opts['namespace_pkg'] = ".".join([opts['namespace'][-1],
-                                          opts['package']])
-    else:
-        opts['root_pkg'] = opts['package']
-        opts['namespace_pkg'] = opts['package']
-    if opts['update']:
-        if not os.path.exists(project_name):
-            raise DirectoryDoesNotExist(
-                "Project {project} does not exist and thus cannot be "
-                "updated!".format(project=project_name))
-        opts = info.project(opts)
-        # Reset project name since the one from setup.cfg might be different
-        opts['project'] = project_name
-    return opts
-
-
-def _init_git(scaffold):
-    """Add revision control to the generated files."""
-    opts = scaffold.options
-    proj_struct = scaffold.changed_structure
-    if not opts['update'] and not repo.is_git_repo(opts['project']):
-        repo.init_commit_repo(opts['project'], proj_struct)
-
-
-def create_project(opts):
-    """Create the project's directory structure
-
-    Args:
-        opts (dict): options of the project
-    """
-    scaffold = Scaffold(opts, make_structure(opts),
-                        before_generate=[_verify_options_consistency],
-                        after_generate=[_init_git])
-
-    # Activate the extensions
-    extensions = opts.get('extensions', [])
-    for extend in extensions:
-        extend(scaffold)
-
-    # Call the before_generate hooks
-    for hook in scaffold.before_generate:
-        hook(scaffold)
-
-    # Decide which files should be generated, and do the job
-    proj_struct = apply_update_rules(scaffold.structure, scaffold.options)
-    proj_struct = add_namespace(opts, proj_struct)
-    # ^ add namespace here, so extensions may benefit
-    changed = create_structure(proj_struct,
-                               update=opts['update'] or opts['force'])
-    scaffold.changed_structure = changed
-
-    # Call the before_generate hooks
-    for hook in scaffold.after_generate:
-        hook(scaffold)
-
-
-def _verify_options_consistency(scaffold):
-    """Perform some sanity checks about the given options."""
-    opts = scaffold.options
-    if os.path.exists(opts['project']):
-        if not opts['update'] and not opts['force']:
-            raise DirectoryAlreadyExists(
-                "Directory {dir} already exists! Use the `update` option to "
-                "update an existing project or the `force` option to "
-                "overwrite an existing directory.".format(dir=opts['project']))
-    if not utils.is_valid_identifier(opts['package']):
-        raise InvalidIdentifier(
-            "Package name {} is not a valid "
-            "identifier.".format(opts['package']))
-
-
 def main(args):
     """PyScaffold is a tool for putting up the scaffold of a Python project.
     """
     opts = parse_args(args)
-    opts = get_default_opts(opts['project'], **opts)
-    create_project(opts)
+    opts = api.get_default_opts(opts['project'], **opts)
+    api.create_project(opts)
     if opts['update'] and not opts['force']:
         note = "Update accomplished!\n" \
                "Please check if your setup.cfg still complies with:\n" \
