+++ tests/extensions/test_cookiecutter.py
@@ -5,7 +5,8 @@ from os.path import exists as path_exists
 
 import pytest
 
-from pyscaffold.cli import create_project, get_default_opts, run
+from pyscaffold.api import create_project, get_default_opts
+from pyscaffold.cli import run
 from pyscaffold.extensions import cookiecutter
 from pyscaffold.templates import setup_py
