+++ tests/extensions/test_tox.py
@@ -3,7 +3,8 @@
 import sys
 from os.path import exists as path_exists
 
-from pyscaffold.cli import create_project, get_default_opts, run
+from pyscaffold.api import create_project, get_default_opts
+from pyscaffold.cli import run
 from pyscaffold.extensions import tox
 
 __author__ = "Anderson Bravalheri"
